+++ metadata.rb
@@ -1,8 +1,8 @@
-name             'java-windows'
+name             'java_windows'
 maintainer       'Lonnie VanZandt'
 maintainer_email 'lonniev@gmail.com'
 license          'All rights reserved'
-description      'Installs/Configures java-windows'
+description      'Installs/Configures java_windows'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'