+++ recipes/default.rb
@@ -1,5 +1,5 @@
 #
-# Cookbook Name:: java-windows
+# Cookbook Name:: java_windows
 # Recipe:: default
 #
 # Copyright 2015, Lonnie VanZandt
@@ -7,11 +7,11 @@
 # All rights reserved - Do Not Redistribute
 #
 
-java-windows_jvm node['java-windows']['title'] do
+java_windows_jvm node['java_windows']['title'] do
   action :install
   
-  jvm node['java-windows']['jvm']
-  version node['java-windows']['version']
+  jvm node['java_windows']['jvm']
+  version node['java_windows']['version']
   update_path true
   set_home true
 end